% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_additional_functions.R
\name{add_qb_epa}
\alias{add_qb_epa}
\title{Compute QB epa}
\usage{
add_qb_epa(d)
}
\arguments{
\item{d}{is a Data frame of play-by-play data scraped using \code{\link{fast_scraper}}.}
}
\description{
Compute QB epa
}
\details{
Add the variable 'qb_epa', which gives QB credit for EPA for up to the point where
a receiver lost a fumble after a completed catch and makes EPA work more
like passing yards on plays with fumbles
}
