% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_nflfastR_pbp.R
\name{build_nflfastR_pbp}
\alias{build_nflfastR_pbp}
\title{Build a complete nflfastR data set}
\usage{
build_nflfastR_pbp(
  game_ids,
  source = "nfl",
  pp = FALSE,
  ...,
  decode = FALSE,
  rules = TRUE
)
}
\arguments{
\item{game_ids}{Vector of character ids (see details for further information).}

\item{source}{Character - \code{nfl} for the NFL.com page or \code{old} for the old gamecenter. For \code{old}, old_game_id must be supplied}

\item{pp}{Logical - either \code{TRUE} or \code{FALSE} (see details for further information)}

\item{...}{Additional arguments passed to the scraping functions (for internal use)}

\item{decode}{If \code{TRUE}, the function \code{\link{decode_player_ids}} will be executed.}

\item{rules}{If \code{FALSE}, printing of the header and footer in the console output will be suppressed.}
}
\value{
An nflfastR play-by-play data frame with like it can be loaded from \url{https://github.com/guga31bb/nflfastR-data}.
}
\description{
`build_nflfastR_pbp()` is a convenient wrapper around 5 nflfastR functions:

* [fast_scraper()]
* [clean_pbp()]
* [add_qb_epa()]
* [add_xyac()]
* [decode_player_ids()]

Please see the documentation of each function to learn about the output.
}
\examples{
\donttest{
# Build nflfastR pbp for the 2018 and 2019 Super Bowls
build_nflfastR_pbp(c("2018_21_NE_LA", "2019_21_SF_KC"))
}
}
