% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_kicking_stats.R
\name{calculate_kicking_stats}
\alias{calculate_kicking_stats}
\title{Summarize Weekly Kicking stats}
\usage{
calculate_kicking_stats(pbp)
}
\arguments{
\item{pbp}{A Data frame of NFL play-by-play data typically loaded with
\code{\link[=load_pbp]{load_pbp()}} or \code{\link[=build_nflfastR_pbp]{build_nflfastR_pbp()}}.}
}
\value{
a dataframe of kicking stats
}
\description{
Build columns that aggregate kicking stats at the game level.
}
\examples{

\donttest{
pbp <- nflreadr::load_pbp(2021)
calculate_kicking_stats(pbp)
}

}
\seealso{
\url{https://nflreadr.nflverse.com/reference/load_player_stats.html} for the nflreadr function to download this from repo (\code{stat_type = "kicking"})
}
