% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_player_stats_gamescript.R
\name{calculate_player_stats_gamescript}
\alias{calculate_player_stats_gamescript}
\title{Get player stats broken down by gamescript.}
\usage{
calculate_player_stats_gamescript(pbp, weekly = FALSE, simple = FALSE)
}
\arguments{
\item{pbp}{A data frame of play-by-play data obtained by running load_pbp().}

\item{weekly}{TRUE/FALSE value indicating if you want your stats broken down by week.}

\item{simple}{Selects only a small amount of columns in the return data frame.}
}
\value{
A data frame.
}
\description{
Get player stats broken down by gamescript.
}
\examples{
pbp_2023 <- nflfastR::load_pbp(2023) \%>\%
  dplyr::filter(week <= 16)
offplayer_pts <- pbp_2023 \%>\%
  nflfastR::calculate_player_stats()
}
