% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_game_stats_kicking.R
\name{calculate_player_stats_kicking}
\alias{calculate_player_stats_kicking}
\title{Summarize Kicking Stats}
\usage{
calculate_player_stats_kicking(...)
}
\arguments{
\item{...}{Leave this empty. The function is deprecated}
}
\value{
Nothing. Function is deprecated.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because we have a new, much better and
harmonized approach in \code{\link[=calculate_stats]{calculate_stats()}}.
}
\examples{
\donttest{
try({# to avoid CRAN test problems
 # calculate_player_stats_kicking()
})
}
}
\keyword{internal}
