% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_series_conversion_rates.R
\name{calculate_series_conversion_rates}
\alias{calculate_series_conversion_rates}
\title{Compute Series Conversion Information from Play by Play}
\usage{
calculate_series_conversion_rates(pbp, weekly = FALSE)
}
\arguments{
\item{pbp}{Play-by-play data as returned by \code{\link[=load_pbp]{load_pbp()}}, \code{\link[=build_nflfastR_pbp]{build_nflfastR_pbp()}}, or
\code{\link[=fast_scraper]{fast_scraper()}}.}

\item{weekly}{If \code{TRUE}, returns week-by-week stats, otherwise,
season-by-season stats in argument \code{pbp}.}
}
\value{
A data frame of series conversion results
}
\description{
A "Series" is a series of downs a team plays to either convert
to a new 1st down (on offense) or prevent the offense from converting a new
1st down (on defense). The series conversion rate describes how many series
have been either converted to a new 1st down or ended in a touchdown.
This function computes series conversion rates on offense and defense from
nflverse play-by-play data along with other series result rates.
}
\examples{
\donttest{
  pbp <- nflfastR::load_pbp(2021)

  weekly <- calculate_series_conversion_rates(pbp, weekly = TRUE)
  dplyr::glimpse(weekly)

  overall <- calculate_series_conversion_rates(pbp, weekly = FALSE)
  dplyr::glimpse(overall)
}
}
