% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_standings.R
\name{calculate_standings}
\alias{calculate_standings}
\title{Compute Division Standings and Conference Seeds from Play by Play}
\usage{
calculate_standings(pbp, tiebreaker_depth = 3, playoff_seeds = NULL)
}
\arguments{
\item{pbp}{nflfastR play-by-play data either loaded via \code{load_pbp()} or
computed via \code{build_nflfastR_pbp()}.}

\item{tiebreaker_depth}{A single value equal to 1, 2, or 3. The default is 3. The
value controls the depth of tiebreakers that shall be applied. The deepest
currently implemented tiebreaker is strength of schedule. The following
values are valid:
\describe{
\item{tiebreaker_depth = 1}{Break all ties with a coinflip. Fastest variant.}
\item{tiebreaker_depth = 2}{Apply head-to-head and division win percentage tiebreakers. Random if still tied.}
\item{tiebreaker_depth = 3}{Apply all tiebreakers through strength of schedule. Random if still tied.}
}}

\item{playoff_seeds}{Number of playoff teams per conference. If \code{NULL} (the
default), the function will try to split \code{pbp} in seasons prior 2020 (6
seeds) and 2020ff (7 seeds). If set to a numeric, it will be used for all
seasons in \code{pbp}!}
}
\value{
A tibble with NFL regular season standings
}
\description{
Compute Division Standings and Conference Seeds from Play by Play
}
\examples{
\donttest{
try({# to avoid CRAN test problems
  pbp <- nflreadr::load_pbp(c(2018, 2021))
  calculate_standings(pbp)
})
}
}
