% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.datatable.aware}
\alias{.datatable.aware}
\title{data.table}
\format{
An object of class \code{logical} of length 1.
}
\usage{
.datatable.aware
}
\description{
data.table
}
\keyword{datasets}
