% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-level_scraper.R
\name{fast_scraper}
\alias{fast_scraper}
\title{Get NFL Play by Play Data}
\usage{
fast_scraper(game_ids, source = "nfl", pp = FALSE, ..., in_builder = FALSE)
}
\arguments{
\item{game_ids}{Vector of character ids (see details for further information).}

\item{source}{Character - \code{nfl} for the NFL.com page or \code{old} for the old gamecenter. For \code{old}, old_game_id must be supplied}

\item{pp}{Logical - either \code{TRUE} or \code{FALSE} (see details for further information)}

\item{...}{Additional arguments passed to the scraping functions (for internal use)}

\item{in_builder}{If \code{TRUE}, the final message will be suppressed (for usage inside of \code{\link{build_nflfastR_pbp}}).}
}
\value{
Data frame where each individual row represents a single play for
all passed game_ids containing the following
}
\description{
Get NFL Play by Play Data
}
\details{
To load valid game_ids please use the package function \code{\link{fast_scraper_schedules}}.

The \code{source} parameter controls from which source the data is being
loaded. The old parameters \code{rs} as well as \code{gc}
are not valid anymore. Please use \code{nfl} or \code{old}.

The \code{pp} parameter controls if the scraper should use parallel processing.
Please note that the initiating process takes a few seconds which means it
may be better to set \code{pp = FALSE} if you are scraping just a few games.
}
\examples{
\donttest{
# Get pbp data for two games
fast_scraper(c("2019_01_GB_CHI", "2013_21_SEA_DEN"))
}
}
\seealso{
\code{\link{field_descriptions}}, \url{https://nflfastr.com/articles/field_descriptions.html}
}
