% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-level_scraper.R
\name{fast_scraper_clips}
\alias{fast_scraper_clips}
\title{Get NFL Play by Play Highlight Clips}
\usage{
fast_scraper_clips(game_ids, pp = FALSE)
}
\arguments{
\item{game_ids}{Vector of numeric or character ids}

\item{pp}{Logical - either \code{TRUE} or \code{FALSE} (see details for further information)}
}
\value{
Data frame containing game_id, play_id for all plays with available
highlightclip and the clip url
}
\description{
Get NFL Play by Play Highlight Clips
}
\details{
To load valid game_ids please use the package function \code{\link{fast_scraper_schedules}}.
The \code{pp} parameter controls if the scraper should use parallel processing.
Please note that the initiating process takes a few seconds which means it
may be better to set \code{pp = FALSE} if you are scraping just a few games.
}
\examples{

# Get highlight clips for two 2019 games using parallel processing
# game_ids <- c("2019090804", "2019101700")
# clips <- fast_scraper_clips(game_ids, pp = TRUE)
}
