% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-level_scraper.R
\name{fast_scraper_roster}
\alias{fast_scraper_roster}
\title{Get team rosters for multiple seasons}
\usage{
fast_scraper_roster(seasons, pp = FALSE)
}
\arguments{
\item{seasons}{A vector of 4-digit years associated with given NFL seasons}

\item{pp}{Logical - either \code{TRUE} or \code{FALSE} (see details for further information)}
}
\value{
Data frame where each individual row represents a player in
the roster of the given team and season containing the following information:
\describe{
\item{season}{}
\item{team}{}
\item{position}{}
\item{depth_chart_position}{}
\item{jersey_number}{}
\item{status}{}
\item{full_name}{}
\item{first_name}{}
\item{last_name}{}
\item{birth_date}{}
\item{height}{}
\item{weight}{}
\item{college}{}
\item{high_school}{}
\item{gsis_id}{}
\item{espn_id}{}
\item{sportradar_id}{}
\item{yahoo_id}{}
\item{rotowire_id}{}
\item{update_dt}{}
\item{headshot_url}{}
}
}
\description{
Given years return a dataset with each player listed as part of the roster.
}
\details{
The roster data is accessed via the free to use Sleeper API.
The \code{pp} parameter controls if the scraper should use parallel processing.
Please note that the initiating process takes a few seconds which means it
may be better to set \code{pp = FALSE} if you are scraping just a few seasons.
}
\examples{
\donttest{
# Roster of the 2019 and 2020 seasons
fast_scraper_roster(2019:2020)
}
}
