% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-level_scraper.R
\name{fast_scraper_schedules}
\alias{fast_scraper_schedules}
\title{Get NFL Season Schedules}
\usage{
fast_scraper_schedules(seasons, pp = FALSE)
}
\arguments{
\item{seasons}{Vector of numeric or character 4 digit seasons}

\item{pp}{Logical - either \code{TRUE} or \code{FALSE} (see details for further information)}
}
\value{
Data frame containing the follwoing detailed game information:
\itemize{
\item{season} - 4 digit season year.
\item{season_type} - Either 'PRE', 'REG', 'POST', 'PRO'.
\item{week} - Numeric week number.
\item{game_id} - Unique game identifier.
\item{alt_game_id} - Alternative and much more intuitive identifier introduced by Lee Sharpe and set to \code{NA} for Hall of Fame Week, Preseason and Pro Bowl
\item{game_date} - Game date in format dd/mm/yyyy.
\item{game_time_eastern} - Kickoff time in eastern time zone.
\item{game_time_local} - Kickoff time in local time zone.
\item{iso_time} - Kickoff time according ISO 8601.
\item{home_team} - Home team abbreviation
\item{away_team} - Away team abbreviation
\item{home_team_name} - Home team full name
\item{away_team_name} - Away team full name
\item{home_nickname} - Home team nick name
\item{away_nickname} - Away team nick name
\item{home_team_id} - Home team id (can be used with the package function \code{\link{fast_scraper_roster}}).
\item{away_team_id} - Away team id (can be used with the package function \code{\link{fast_scraper_roster}}).
\item{game_type} - One of 'REG', 'WC', 'DIV', 'CON', 'SB' indicating if a game was a regular season game or one of the playoff rounds.
\item{week_name} - Full description of week
\item{site_city} - Game site city.
\item{site_fullname} - Game site name.
\item{site_state} - Game site state.
\item{site_roof_type} - Game site roof type.
\item{network_channel} - Name of broadcasting network channel.
}
}
\description{
Get NFL Season Schedules
}
\details{
The \code{pp} parameter controls if the scraper should use parallel processing.
Please note that the initiating process takes a few seconds which means it
may be better to set \code{pp = FALSE} if you are scraping less than 10 seasons.
}
\examples{

# Get schedules for the whole 2015 - 2018 seasons
# seasons <- 2015:2018
# schedules <- fast_scraper_schedules(seasons)
}
