% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{field_descriptions}
\alias{field_descriptions}
\title{nflfastR Field Descriptions}
\format{
A data frame including names and descriptions of all variables in
an nflfastR dataset.
}
\usage{
field_descriptions
}
\description{
nflfastR Field Descriptions
}
\examples{
\donttest{
field_descriptions
}
}
\seealso{
The searchable table on the
\href{https://nflfastr.com/articles/field_descriptions.html}{nflfastR website}
}
\keyword{datasets}
