% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_game_stats.R
\name{get_player_stats}
\alias{get_player_stats}
\title{Get official game stats}
\usage{
get_player_stats(df, weekly = TRUE)
}
\arguments{
\item{df}{is a Data frame of play-by-play data scraped using \code{\link[=fast_scraper]{fast_scraper()}}.}

\item{weekly}{Defaults to TRUE. If true, returns week-by-week stats, otherwise, stats
for the entire dataframe.}
}
\value{
The following columns
\describe{
\item{player_id}{ID of the player. Use this to join to other sources.}
\item{player_name}{Name of the player}
\item{season}{Season if \code{weekly} is TRUE}
\item{week}{Week if \code{weekly} is TRUE}
\item{other_things}{fill this out}
}
}
\description{
Get official game stats
}
\details{
Build columns that aggregate official passing, rushing, and receiving stats
at the game level or at the level of the entire df passed.
}
