% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_multi_laterals.R
\name{load_multi_laterals}
\alias{load_multi_laterals}
\title{Load and aggregate multi-lateral plays}
\usage{
load_multi_laterals()
}
\value{
A data frame aggregating yards gained per player per multi-lateral play on record.
}
\description{
Load and aggregate multi-lateral plays
}
\examples{
\donttest{
mult_lats <- load_multi_laterals()
}
}
