% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{nfl_stats_variables}
\alias{nfl_stats_variables}
\title{NFL Stats Variables}
\format{
A data frame explaining all variables returned by the function
\code{\link[=calculate_stats]{calculate_stats()}}.
}
\usage{
nfl_stats_variables
}
\description{
NFL Stats Variables
}
\examples{
\donttest{
nfl_stats_variables
}
}
\keyword{datasets}
