% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nflverse_sitrep}
\alias{nflverse_sitrep}
\title{nflverse sitrep}
\usage{
nflverse_sitrep()
}
\value{
Situation report of R and package/dependencies.
}
\description{
See \code{nflreadr::nflverse_sitrep} for details.
}
\keyword{internal}
