% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflfastR-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{load_pbp}
\alias{load_player_stats}
\alias{load_team_stats}
\alias{load_schedules}
\alias{load_rosters}
\alias{nflverse_sitrep}
\alias{most_recent_season}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{nflreadr}{\code{\link[nflreadr:load_pbp]{load_pbp()}}, \code{\link[nflreadr:load_player_stats]{load_player_stats()}}, \code{\link[nflreadr:load_rosters]{load_rosters()}}, \code{\link[nflreadr:load_schedules]{load_schedules()}}, \code{\link[nflreadr:load_team_stats]{load_team_stats()}}, \code{\link[nflreadr:most_recent_season]{most_recent_season()}}, \code{\link[nflreadr:nflverse_sitrep]{nflverse_sitrep()}}}
}}

