% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{stat_ids}
\alias{stat_ids}
\title{NFL Stat IDs and their Meanings}
\format{
A data frame including NFL stat IDs, names and descriptions used in
an nflfastR dataset.
}
\source{
\url{https://www.nflgsis.com/gsis/Documentation/Partners/StatIDs.html}
}
\usage{
stat_ids
}
\description{
NFL Stat IDs and their Meanings
}
\examples{
\donttest{
stat_ids
}
}
\keyword{datasets}
