% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_database_functions.R
\name{update_db}
\alias{update_db}
\title{Update or create a nflfastR play by play database}
\usage{
update_db(
  dbdir = ".",
  dbname = "pbp_db",
  tblname = "nflfastR_pbp",
  force_rebuild = FALSE,
  db_connection = NULL
)
}
\arguments{
\item{dbdir}{Directory in which the database is or shall be located}

\item{dbname}{File name of an existing or desired SQLite database within \code{dbdir}}

\item{tblname}{The name of the play by play data table within the database}

\item{force_rebuild}{Hybrid parameter (logical or numeric) to rebuild parts
of or the complete play by play data table within the database (please see details for further information)}

\item{db_connection}{A \code{DBIConnection} object, as returned by
\code{\link[DBI]{dbConnect}} (please see details for further information)}
}
\description{
\code{update_db} updates or creates a SQLite database with \code{nflfastR}
play by play data of all completed games since 1999.
}
\details{
This function creates and updates a data table with the name \code{tblname}
within a SQLite database located in \code{dbdir} and named \code{dbname}.
The data table combines all play by play data for every available game back
to the 1999 season and adds the most recent completed games as soon as they
are available for \code{nflfastR}.

The argument \code{force_rebuild} is of hybrid type. It can rebuild the play
by play data table either for the whole nflfastR era (with \code{force_rebuild = TRUE})
or just for specified seasons (e.g. \code{force_rebuild = c(2018, 2020)}).
The latter is intended to be used for running seasons because the NFL fixes
bugs in the play by play data during the week and it is recommended to rebuild
the current season every Wednesday during the season.

The parameter \code{db_connection} is intended for advanced users who want
to use other DBI drivers, such as MariaDB, Postgres or odbc. Please note that
the arguments \code{dbdir} and \code{dbname} are dropped in case a \code{db_connection}
is provided but the argument \code{tblname} will still be used to write the
data table into the database.
}
