% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_decode_player_ids.R
\name{decode_player_ids}
\alias{decode_player_ids}
\title{Decode the player IDs in nflfastR play-by-ply data}
\usage{
decode_player_ids(pbp, ...)
}
\arguments{
\item{pbp}{is a Data frame of play-by-play data scraped using \code{\link{fast_scraper}}.}

\item{...}{Additional arguments passed to a message function (for internal use).}
}
\value{
The input Data Frame of the parameter 'pbp' with decoded player IDs.
}
\description{
Decode the player IDs in nflfastR play-by-ply data
}
\details{
Take all columns ending with \code{player_id} as well as
\code{passer_id}, \code{rusher_id}, \code{receiver_id}, \code{id} of an
nflfastR play-by-play Data set and decode the player IDs to the commonly
known GSIS ID format 00-00xxxxx.
The function requires the package \code{furrr} if the data frame
\code{pbp} has more than 4500 rows.
}
\examples{
\donttest{
# Decode IDs by Patrick Mahomes and Julio Jones
decode_player_ids(data.frame(
  name = c("P.Mahomes", "J.Jones"),
  id = c(
    "32013030-2d30-3033-3338-3733fa30c4fa",
    "32013030-2d30-3032-3739-3434d4d3846d"
  )
))
}
}
