% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-level_scraper.R
\name{fast_scraper_roster}
\alias{fast_scraper_roster}
\title{Get team rosters for multiple seasons and teams}
\usage{
fast_scraper_roster(team_ids, seasons, pp = FALSE)
}
\arguments{
\item{team_ids}{A string vector containing the IDs for NFL Team(s)
(see details for more information)}

\item{seasons}{A string vector of 4-digit years associated with given NFL seasons}

\item{pp}{Logical - either \code{TRUE} or \code{FALSE} (see details for further information)}
}
\value{
Data frame where each individual row represents a player in
the roster of the given team and season listed by the NFL
containing the following information:
\itemize{
\item{team.season}
\item{teamPlayers.displayName}
\item{teamPlayers.firstName}
\item{teamPlayers.middleName}
\item{teamPlayers.lastName}
\item{teamPlayers.suffix}
\item{teamPlayers.status}
\item{teamPlayers.position}
\item{teamPlayers.positionGroup}
\item{teamPlayers.nflId}
\item{teamPlayers.esbId}
\item{teamPlayers.gsisId}
\item{teamPlayers.birthDate}
\item{teamPlayers.homeTown}
\item{teamPlayers.collegeId}
\item{teamPlayers.collegeName}
\item{teamPlayers.jerseyNumber}
\item{teamPlayers.height}
\item{teamPlayers.weight}
\item{team.teamId}
\item{team.abbr}
\item{team.cityState}
\item{team.fullName}
\item{team.nick}
\item{team.conferenceAbbr}
\item{team.divisionAbbr}
\item{teamPlayers.headshot_url}
\item{teamPlayers.profile_url}
}
}
\description{
Given team_ids and years, return a dataset with each
player the NFL has listed as part of the roster.
}
\details{
To find team associated Team IDs use the \code{\link{teams_colors_logos}}
dataset stored in this package!
The \code{pp} parameter controls if the scraper should use parallel processing.
Please note that the initiating process takes a few seconds which means it
may be better to set \code{pp = FALSE} if you are scraping just a few teams/seasons.
}
\examples{
# Roster of Steelers in 2018, no parallel processing
# rosters <- fast_scraper_roster("3900", 2018, pp = FALSE)

# Roster of Steelers and Seahawks in 2016 & 2019 using parallel processing
# rosters <- fast_scraper_roster(c("3900", "4600"), c("2016", "2019"), pp = TRUE)
}
