% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-level_scraper.R
\name{fast_scraper_roster}
\alias{fast_scraper_roster}
\title{Load Team Rosters for Multiple Seasons}
\usage{
fast_scraper_roster(seasons = most_recent_season(roster = TRUE))
}
\arguments{
\item{seasons}{a numeric vector of seasons to return, defaults to returning
this year's data if it is March or later. If set to \code{TRUE}, will return all available data.}
}
\value{
A tibble of season-level roster data.
}
\description{
Load Rosters
}
\examples{
\donttest{
# Roster of the 2019 and 2020 seasons
fast_scraper_roster(2019:2020)
}
}
\seealso{
For information on parallel processing and progress updates please
see \link{nflfastR}.
}
