% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-level_scraper.R
\name{fast_scraper_roster}
\alias{fast_scraper_roster}
\title{Load Team Rosters for Multiple Seasons}
\usage{
fast_scraper_roster(...)
}
\value{
A tibble of season-level roster data.
}
\description{
Load Rosters
}
\details{
See \code{\link[nflreadr:load_rosters]{nflreadr::load_rosters}} for details.
}
\examples{
\donttest{
# Roster of the 2019 and 2020 seasons
try({# to avoid CRAN test problems
fast_scraper_roster(2019:2020)
})
}
}
\seealso{
For information on parallel processing and progress updates please
see \link{nflfastR}.
}
