% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-level_scraper.R
\name{fast_scraper_schedules}
\alias{fast_scraper_schedules}
\title{Load NFL Season Schedules}
\description{
Load NFL Season Schedules
}
\details{
See \code{\link[nflreadr:load_schedules]{nflreadr::load_schedules}} for details.
}
\examples{
\donttest{
# Get schedules for the whole 2015 - 2018 seasons
fast_scraper_schedules(2015:2018)
}
}
\seealso{
For information on parallel processing and progress updates please
see \link{nflfastR}.
}
