% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-level_scraper.R
\name{fast_scraper_schedules}
\alias{fast_scraper_schedules}
\title{Load NFL Season Schedules}
\usage{
fast_scraper_schedules(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[nflreadr:load_schedules]{nflreadr::load_schedules}}
  \describe{
    \item{\code{seasons}}{a numeric vector of seasons to return, default \code{TRUE} returns all available data.}
  }}
}
\value{
A tibble of game information for past and/or future games.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated. Please use \code{\link[nflreadr:load_schedules]{nflreadr::load_schedules}}.
}
\details{
See \code{\link[nflreadr:load_schedules]{nflreadr::load_schedules}} for details.
}
\examples{
\donttest{
# Get schedules for the whole 2015 - 2018 seasons
try({# to avoid CRAN test problems
# fast_scraper_schedules(2015:2018)
})
}
}
\seealso{
For information on parallel processing and progress updates please
see \link{nflfastR}.
}
\keyword{internal}
