% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{field_descriptions}
\alias{field_descriptions}
\title{PBP column descriptions}
\format{
A dataframe with two columns: field, and description.

\describe{
 \item{field}{Name of column as shown in PBP dataframe}
 \item{description}{Description of column}
}
}
\usage{
field_descriptions
}
\description{
This dataframe describes the columns of the play-by-play data created by the \code{build_nflfastr_pbp()} function.
}
\keyword{datasets}
