% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_game_stats.R
\name{get_player_stats}
\alias{get_player_stats}
\title{Get Official Game Stats}
\usage{
get_player_stats(pbp, weekly = TRUE)
}
\arguments{
\item{pbp}{A Data frame of NFL play-by-play data \strong{with decoded player IDs}
typically loaded with \code{\link[=load_pbp]{load_pbp()}} or \code{\link[=build_nflfastR_pbp]{build_nflfastR_pbp()}}.
For ID decoding please see \code{\link[=decode_player_ids]{decode_player_ids()}}.}

\item{weekly}{If \code{TRUE}, returns week-by-week stats, otherwise, stats
for the entire Data frame.}
}
\value{
A data frame including the following columns:
\describe{
\item{player_id}{ID of the player. Use this to join to other sources.}
\item{player_name}{Name of the player}
\item{season}{Season if \code{weekly} is \code{TRUE}}
\item{week}{Week if \code{weekly} is \code{TRUE}}
\item{cmp}{The number of completed passes.}
\item{att}{The number of pass attempts as defined by the NFL.}
\item{passing_yards}{Yards gained on pass plays.}
\item{pass_tds}{The number of passing touchdowns.}
\item{ints}{The number of interceptions thrown.}
\item{fumble_lost_sack}{The number of sacks with a lost fumble.}
\item{carries}{The number of rush attempts (incl. scrambles). Rushes after
a lateral reception don't count as carry.}
\item{rushing_yards}{Yards gained when rushing with the ball (incl. scrambles).
Also includes yards gained after receiving a lateral on a play that started
with a rushing attempt.}
\item{tds_rush}{The number of rushing touchdowns (incl. scrambles).
Also includes touchdowns after receiving a lateral on a play that started
with a rushing attempt.}
\item{fumble_lost_rush}{The number of rushes with a lost fumble.}
\item{rec}{The number of pass receptions. Lateral receptions officially
don't count as reception.}
\item{tgt}{The number of pass plays where the player was the targeted receiver.}
\item{receiving_yards}{Yards gained after a pass reception. Includes yards
gained after receiving a lateral on a play that started as a pass play.}
\item{tds_rec}{The number of touchdowns following a pass reception.
Also includes touchdowns after receiving a lateral on a play that started
as a pass play.}
\item{fumble_lost_rec}{The number of fumbles after a pass reception.}
}
}
\description{
Build columns that aggregate official passing, rushing, and receiving stats
either at the game level or at the level of the entire data frame passed.
}
\examples{
\donttest{
pbp <- nflfastR::load_pbp(2020)

weekly <- get_player_stats(pbp)
dplyr::glimpse(weekly)

overall <- get_player_stats(pbp, weekly = FALSE)
dplyr::glimpse(overall)
}
}
