% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflfastR-package.R
\docType{package}
\name{nflfastR-package}
\alias{nflfastR}
\alias{nflfastR-package}
\title{nflfastR: Functions to Efficiently Access NFL Play by Play Data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A set of functions to access National Football
    League play-by-play data from <https://www.nfl.com/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.nflfastr.com/}
  \item \url{https://github.com/mrcaseb/nflfastR}
  \item Report bugs at \url{https://github.com/mrcaseb/nflfastR/issues}
}

}
\author{
\strong{Maintainer}: Ben Baldwin \email{bbaldwin206@gmail.com}

Authors:
\itemize{
  \item Sebastian Carl \email{mrcaseb@gmail.com}
}

Other contributors:
\itemize{
  \item Lee Sharpe [contributor]
  \item Maksim Horowitz \email{maksim.horowitz@gmail.com} [contributor]
  \item Ron Yurko \email{ryurko@stat.cmu.edu} [contributor]
  \item Samuel Ventura \email{samventura22@gmail.com} [contributor]
  \item Tan Ho [contributor]
}

}
\keyword{internal}
