% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflfastR-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{load_pbp}
\alias{load_player_stats}
\alias{load_schedules}
\alias{load_rosters}
\alias{nflverse_sitrep}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{nflreadr}{\code{\link[nflreadr]{load_pbp}}, \code{\link[nflreadr]{load_player_stats}}, \code{\link[nflreadr]{load_rosters}}, \code{\link[nflreadr]{load_schedules}}, \code{\link[nflreadr:sitrep]{nflverse_sitrep}}}
}}

