% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Get a Situation Report on System, nflverse Package Versions and Dependencies}
\description{
This function gives a quick overview of the versions of R and
the operating system as well as the versions of nflverse packages, options,
and their dependencies. It's primarily designed to help you get a quick
idea of what's going on when you're helping someone else debug a problem.
}
\details{
See \code{\link[nflreadr:sitrep]{nflreadr::nflverse_sitrep}} for details.
}
\examples{
\donttest{
report(recursive = FALSE)
report(pkg = "nflreadr", recursive = TRUE)
}
}
